refactor: Regional UI와 runtime lifecycle 분리
