+++ tools/check_frontend.ps1
@@ -44,6 +44,11 @@ try {
         throw "Frontend Regional pure data smoke failed with exit code $LASTEXITCODE."
     }
 
+    & node "tests\frontend_regional_runtime_smoke.mjs"
+    if ($LASTEXITCODE -ne 0) {
+        throw "Frontend Regional runtime lifecycle smoke failed with exit code $LASTEXITCODE."
+    }
+
     & npx --yes --package "typescript@$TypeScriptVersion" -- tsc -p jsconfig.json
     if ($LASTEXITCODE -ne 0) {
         throw "TypeScript check failed with exit code $LASTEXITCODE."
